Rename fiat module to price, move formatAmount from constants
