+++ src/utils/constants.ts
@@ -15,27 +15,6 @@ export const TXS_PER_PAGE = 25;
 export const UX_DELAY_FAST = 1000;
 export const UX_DELAY_NORMAL = 2000;
 export const UX_DELAY_SLOW = 3000;
-/**
- * Format a crypto amount with commas and smart decimal reduction for large numbers.
- * e.g. 1234.56789 -> "1,234.56789"
- * e.g. 1234567.89 -> "1,234,567.89"
- * e.g. 1234567890 -> "1,234,567,890" (0 decimals for billions)
- */
-export function formatAmount(value: number): string {
-  if (value === 0) return '0';
-
-  let decimals = 8;
-  if (value >= 1_000_000_000) decimals = 0;
-  else if (value >= 1_000_000) decimals = 2;
-
-  // Use Intl.NumberFormat for reliable comma separators
-  const formatter = new Intl.NumberFormat('en-US', {
-    minimumFractionDigits: 0,
-    maximumFractionDigits: decimals
-  });
-
-  return formatter.format(value);
-}
 
 /**
  * Truncates a long ID (like a TXID or address) for display.
